move `migrate-prettier` to `codemods/config/migrate-prettier`
