+++ packages/@tailwindcss-upgrade/src/index.ts
@@ -6,6 +6,7 @@ import path from 'node:path'
 import postcss from 'postcss'
 import { migrateJsConfig } from './codemods/config/migrate-js-config'
 import { migratePostCSSConfig } from './codemods/config/migrate-postcss'
+import { migratePrettierPlugin } from './codemods/config/migrate-prettier'
 import { analyze as analyzeStylesheets } from './codemods/css/analyze'
 import { formatNodes } from './codemods/css/format-nodes'
 import { linkConfigs as linkConfigsToStylesheets } from './codemods/css/link'
@@ -15,7 +16,6 @@ import { split as splitStylesheets } from './codemods/css/split'
 import { migrate as migrateTemplate } from './codemods/template/migrate'
 import { prepareConfig } from './codemods/template/prepare-config'
 import { help } from './commands/help'
-import { migratePrettierPlugin } from './migrate-prettier'
 import { Stylesheet } from './stylesheet'
 import { args, type Arg } from './utils/args'
 import { isRepoDirty } from './utils/git'
